fix teams test
