+++ pipeline/Jenkinsfile
@@ -73,7 +73,7 @@ parameters
                     py.test -k 'smoke' --reruns 1 --ignore=core --disable-pytest-warnings --alluredir=ui_automation_tests/allure-results
                     if [ \$? -eq 0 ]
                     then
-                         py.test -k 'regression' -n '${env.NO_OF_PARALLEL_RUNNERS}' --dist=loadfile --reruns 2 --ignore=core --disable-pytest-warnings --alluredir=ui_automation_tests/allure-results
+                         py.test -k 'regression' -n '${env.NO_OF_PARALLEL_RUNNERS}' --dist=loadfx2x2ile --reruns 2 --ignore=core --disable-pytest-warnings --alluredir=ui_automation_tests/allure-results
                     fi
                     """
                 }
